Add mixins for taxonomy cached properties and promote panels
